requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
PyYAML==6.0.2
httpx==0.27.2
openai==1.47.0